fix certificate generation for isolation segments
Certificate generation was failing because the JSON payload being passed
in to Opsman was not correct. The JSON object needs to be an array even
if of size 1. Without this fix the om-linux call will respond with "could not
execute 'curl' server responded with an error."

This solution was borrowed from:

https://github.com/pivotal-cf/pcf-pipelines/blob/master/functions/generate_cert.sh